moves .../evolutionaryAlgorithm/RuleApplicationsOfPlans.java to .../rewriting/PlanRewritingUtils.java
